refactor: consolidate helpers and simplify project initialization
