+++ packages/create-app/README.md
@@ -1,4 +1,5 @@
 # create-ckb-js-vm-app
+
 create-ckb-js-vm-app is a CLI tool that helps you quickly bootstrap CKB script powered by [ckb-js-vm](https://github.com/nervosnetwork/ckb-js-vm). It allows you to generate a on-chain script project with support for TypeScript.
 
 # Getting Started
